+++ js/util.js
@@ -1,0 +1,5 @@
+const _ = window._;
+
+const getRandomInt = (min, max) => _.random(min, max);
+
+export default getRandomInt;
